OPCODE OVERHAUL

+++ src/stack.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Stack<const N: usize = 32> {
-    inner: [u16; N],
+    inner: [usize; N],
     last: usize,
 }
 
@@ -11,12 +11,12 @@ impl<const N: usize> Stack<N> {
             last: 0,
         }
     }
-    pub fn push(&mut self, element: u16) {
+    pub fn push(&mut self, element: usize) {
         self.last += 1;
         self.inner[self.last] = element;
     }
 
-    pub fn pop(&mut self) -> u16 {
+    pub fn pop(&mut self) -> usize {
         let value = self.inner[self.last];
         self.inner[self.last] = 0;
         self.last -= 1;
